added users can only delete their own threads feature

+++ app/Http/Controllers/ThreadsController.php
@@ -87,6 +87,11 @@ class ThreadsController extends Controller
 
     public function destroy(Channel $channel, Thread $thread)
     {
+        if ($thread->user_id != auth()->id()) {
+            abort(403, 'You do not have permission to do this.');
+        }
+
+
         $thread->delete();
 
         if (request()->wantsJson()) {
